# backend/swagger.yaml
swagger: "2.0"
info:
  title: KnightWise
  description: |
    API documentation for KnightWise, the computer science guided learning platform.

    ## Question Types and Grading
    
    KnightWise supports five non-programming question types, each with its own grading algorithm:
    
    ### Multiple Choice
    - **Input Format**: String (e.g., `"Paris"`)
    - **Grading**: Exact match (case-insensitive, trimmed)
    - **Normalized Score**: 1.0 (correct) or 0.0 (incorrect)
    
    ### Fill-In-the-Blanks
    - **Input Format**: String (e.g., `"Paris"`)
    - **Grading**: Exact or fuzzy matching using Levenshtein distance
    - **Normalized Score**: 0.0 to 1.0 based on similarity (1 - distance/maxLength)
    - **Partial Credit**: Yes (>= 50% similarity, API feedback says "Almost correct!")
    
    ### Select-All-That-Apply
    - **Input Format**: Array of strings (e.g., `["Python", "JavaScript"]`)
    - **Grading**: (correct selections - incorrect selections) / total correct options
    - **Normalized Score**: 0.0 to 1.0 based on how many correct selections
    - **Partial Credit**: Yes
    
    ### Ranked Choice
    - **Input Format**: Array of strings in ranked order (e.g., `["First", "Second", "Third"]`)
    - **Grading**: Kendall tau distance (counts pairwise inversions)
    - **Normalized Score**: 0.0 to 1.0, based on 1.0 - (inversions / maxInversions)
    - **Partial Credit**: Yes
    
    ### Drag-and-Drop
...
    - **Grading**: Correct placements / total items
...
    
...
    
...
paths:
...
  /store:
    get:
      tags:
        - Store
      summary: Fetch all store items.
      operationId: getStoreItems
      description: Retrieves metadata for all available store items.
      security:
        - BearerAuth: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/StoreItemsResponse'
        401:
          description: Unauthorized
        500:
          description: Server Error
  /store/{id}/purchase:
    post:
      tags:
        - Store
      summary: Purchase a store item with coins.
      operationId: purchaseItem
      description: Attempts to purchase a store item using the authenticated user's coins. Deducts coins and inserts a purchase record atomically. Item is not equipped on purchase.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the store item to purchase.
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid item ID or insufficient coins
        401:
          description: Unauthorized
        404:
          description: Item Not Found
        409:
          description: Conflict - Item already purchased
        500:
          description: Server Error
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Include the token in the Authorization header as 'Bearer your-token'"
definitions:
...
  StoreItem:
    type: object
    properties:
      ID:
        type: integer
        example: 1
      TYPE:
        type: string
        example: "flair"
        description: "Item type. One of: flair, profile_picture, background"
      COST:
        type: string
        example: "50.00"
        description: "Returned as string due to MySQL DECIMAL serialization."
      NAME:
        type: string
        example: "Greatest of all Time"
      IS_GUILD_ITEM:
        type: integer
        example: 0
        description: "1 if this is a guild store item, 0 if it is a user store item."
  StoreItemsResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/StoreItem'
...
host: www.knightwise.dev
basePath: /api
schemes:
...